Diff cache, secret rules, multi-server request and registry scan config typed
summary: the last backend own-file readers move to records and ModelJson; EngineJson is no longer used by the backend's file reads.

- DiffCache: one DiffCacheEntry (signature, canonical) per host/config, written atomically; an unreadable entry is a miss (the cache is rebuilt from the scanned files). The checkout-to-data-root legacy cache migration is gone.
- Secret rules: SecretRules.Packaged is the embedded secret_rules.json read strictly and compiled once (a bad rule names itself); the mutable rule cache, resource-reader seam and truthiness helpers are gone, so the tests no longer serialise on a shared cache collection.
- Multi-server CLI: stdin is a MultiServerRequest (plans are the GUI's ServerScanPlan) read strictly; each stdout line is {"type": progress|result|error, ...}; the surrogate escape/unescape converters are gone. Enums use the built-in [JsonStringEnumMemberName]; the custom JsonStringEnumMemberConverter is deleted. Baseline priority is int.
- maint selfcheck-multi-server-paths runs typed plans in process and writes a typed report (7/7 on the fixtures); the all_drives_scope scenario is dropped (the command never honoured scope).
- registry-scan emit-config prints a RegistryScanConfig (typed remote targets via RegistryRemoteTarget.Parse); search limits are int.

+++ gui/DriftBuster.Backend/Models/RegistrySearchRequest.cs
@@ -18,10 +18,10 @@ namespace DriftBuster.Backend.Models
         public string[] Patterns { get; set; } = [];
 
         [JsonPropertyName("max_depth")]
-        public long MaxDepth { get; set; } = 12;
+        public int MaxDepth { get; set; } = 12;
 
         [JsonPropertyName("max_hits")]
-        public long MaxHits { get; set; } = 200;
+        public int MaxHits { get; set; } = 200;
 
         [JsonPropertyName("time_budget_s")]
         public double TimeBudgetSeconds { get; set; } = 10.0;
